+++ package.json
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "test": "jest --watchAll",
-    "start": "ts-node-dev --respawn --transpile-only src/drivers/api.ts"
+    "start": "ts-node-dev --respawn --transpile-only src/main_api.ts"
   },
   "keywords": [],
   "author": "",
