+++ src/domain/Ride.ts
