+++ test/validateCpf.test.ts
@@ -1,4 +1,4 @@
-import { validateCpf } from "../src/application/validateCpf";
+import { validateCpf } from "../src/domain/validateCpf";
 
 test.each([
 	"97456321558",
